Split 'types' from 'orangetypes'
