+++ src/orange/OrangeParams.cc
@@ -23,7 +23,7 @@
 #include "corecel/io/StringUtils.hh"
 
 #include "Data.hh"
-#include "Types.hh"
+#include "OrangeTypes.hh"
 #include "construct/OrangeInput.hh"
 #include "detail/UnitInserter.hh"
 #include "univ/detail/LogicStack.hh"
