+++ src/orange/OrangeTrackView.hh
@@ -13,7 +13,7 @@
 #include "corecel/sys/ThreadId.hh"
 
 #include "Data.hh"
-#include "Types.hh"
+#include "OrangeTypes.hh"
 #include "univ/SimpleUnitTracker.hh"
 #include "univ/UniverseTypeTraits.hh"
 #include "univ/detail/Types.hh"
