+++ src/orange/construct/OrangeInput.hh
@@ -13,8 +13,8 @@
 #include "corecel/cont/Label.hh"
 #include "orange/BoundingBox.hh"
 #include "orange/Data.hh"
+#include "orange/OrangeTypes.hh"
 #include "orange/Translator.hh"
-#include "orange/Types.hh"
 
 namespace celeritas
 {
