+++ src/orange/construct/SurfaceInputBuilder.hh
@@ -10,7 +10,7 @@
 #include "corecel/Types.hh"
 #include "corecel/cont/Span.hh"
 #include "orange/Data.hh"
-#include "orange/Types.hh"
+#include "orange/OrangeTypes.hh"
 
 namespace celeritas
 {
