+++ src/orange/detail/UnitIndexer.hh
@@ -9,7 +9,7 @@
 
 #include <vector>
 
-#include "orange/Types.hh"
+#include "orange/OrangeTypes.hh"
 
 namespace celeritas
 {
