+++ src/orange/univ/VolumeView.hh
@@ -11,7 +11,7 @@
 #include "corecel/data/Collection.hh"
 #include "corecel/math/Algorithms.hh"
 #include "orange/Data.hh"
-#include "orange/Types.hh"
+#include "orange/OrangeTypes.hh"
 
 namespace celeritas
 {
