+++ src/orange/univ/detail/Types.hh
@@ -9,7 +9,7 @@
 
 #include "corecel/OpaqueId.hh"
 #include "corecel/cont/Span.hh"
-#include "orange/Types.hh"
+#include "orange/OrangeTypes.hh"
 
 namespace celeritas
 {
